+++ src/MusicBrainz/Entities/Place.php
@@ -38,7 +38,7 @@ class Place extends AbstractEntity implements EntityInterface
 
     private array $data;
 
-    private MusicBrainz $brainz;
+    protected MusicBrainz $brainz;
 
     /**
      * @param array $place
@@ -86,20 +86,21 @@ class Place extends AbstractEntity implements EntityInterface
 
     /**
      * @return array{
-     *      id: string,
-     *      name: string,
-     *      address: string,
-     *      type-id: string,
-     *      type: string,
-     *      area: ?Area[],
-     *      life-span: ?LifeSpan[],
-     *      coordinates: ?Coordinate[],
-     *      disambiguation: string
+     *     id: string,
+     *     name: string,
+     *     address: string,
+     *     type-id: string,
+     *     type: string,
+     *     area: ?Area[],
+     *     life-span: ?LifeSpan[],
+     *     coordinates: ?Coordinate[],
+     *     disambiguation: string,
+     *     data?: array<string, mixed>
      * }
      */
-    public function getProps(): array
+    public function getProps(bool $includeData = false): array
     {
-        return [
+        $results = [
             'id' => $this->id,
             'name' => $this->name,
             'address' => $this->address,
@@ -110,5 +111,11 @@ class Place extends AbstractEntity implements EntityInterface
             'coordinates' => $this->coordinates,
             'disambiguation' => $this->disambiguation,
         ];
+
+        if ($includeData) {
+            $results['data'] = $this->data;
+        }
+
+        return $results;
     }
 }