+++ src/MusicBrainz/Entities/Series.php
@@ -25,7 +25,7 @@ class Series extends AbstractEntity implements EntityInterface
 
     private array $data;
 
-    private MusicBrainz $brainz;
+    protected MusicBrainz $brainz;
 
     /**
      * @param array $series
@@ -73,17 +73,24 @@ class Series extends AbstractEntity implements EntityInterface
      *     name: string,
      *     type-id: string,
      *     type: string,
-     *     disambiguation: string
+     *     disambiguation: string,
+     *     data?: array<string, mixed>
      * }
      */
-    public function getProps(): array
+    public function getProps(bool $includeData = false): array
     {
-        return [
+        $results = [
             'id' => $this->id,
             'name' => $this->name,
             'type-id' => $this->type_id,
             'type' => $this->type,
             'disambiguation' => $this->disambiguation,
         ];
+
+        if ($includeData) {
+            $results['data'] = $this->data;
+        }
+
+        return $results;
     }
 }